+++ assets/SwitchPage.d388b3e4.js
@@ -1,0 +1,43 @@
+var m=Object.defineProperty;var r=(e,a)=>m(e,"name",{value:a,configurable:!0});import{S as n}from"./Switch.120bc4be.js";import{C as h}from"./ComponentInfo.a57005c5.js";import{F as p}from"./FormControlLabel.f99fa179.js";import{d as t,a6 as l,m as i,t as f,a as b}from"./index.ad7ec8f2.js";import"./SwitchBase.29206069.js";import"./useFormControl.c9566a10.js";import"./useControlled.fd4d74b5.js";import"./Link.38c3aff6.js";import"./copyText.c783d453.js";import"./PaperCode.1b5c232b.js";import"./formControlState.3dcd2da7.js";function c(){return t(p,{disabled:!0,get control(){return t(n,{defaultChecked:!0})},label:"Label"})}r(c,"Label");c.code=`import FormControlLabel from "@suid/material/FormControlLabel";
+import Switch from "@suid/material/Switch";
+
+export default function Label() {
+  return (
+    <FormControlLabel
+      disabled
+      control={<Switch defaultChecked />}
+      label="Label"
+    />
+  );
+}
+`;const S=f("<div></div>"),o={inputProps:{"aria-label":"Switch demo"}};function d(){return(()=>{const e=S.cloneNode(!0);return l(e,t(n,i(o,{defaultChecked:!0})),null),l(e,t(n,o),null),l(e,t(n,i(o,{disabled:!0,defaultChecked:!0})),null),l(e,t(n,i(o,{disabled:!0})),null),e})()}r(d,"BasicSwitches");d.code=`import Switch from "@suid/material/Switch";
+
+const label = { inputProps: { "aria-label": "Switch demo" } };
+
+export default function BasicSwitches() {
+  return (
+    <div>
+      <Switch {...label} defaultChecked />
+      <Switch {...label} />
+      <Switch {...label} disabled defaultChecked />
+      <Switch {...label} disabled />
+    </div>
+  );
+}
+`;function u(){const[e,a]=b(!1);return t(n,{get checked(){return e()},onChange:(w,s)=>{a(s)},inputProps:{"aria-label":"controlled"}})}r(u,"ControlledSwitch");u.code=`import Switch from "@suid/material/Switch";
+import { createSignal } from "solid-js";
+
+export default function ControlledSwitch() {
+  const [checked, setChecked] = createSignal(false);
+
+  return (
+    <Switch
+      checked={checked()}
+      onChange={(event, value) => {
+        setChecked(value);
+      }}
+      inputProps={{ "aria-label": "controlled" }}
+    />
+  );
+}
+`;function j(){return t(h,{get name(){return n.name},docsName:"switches",examples:[d,u,c]})}r(j,"SwitchPage");export{j as default};